+++ src/data-values/index.js
@@ -3,6 +3,7 @@
 
 export * from '#x/BaseConverter';
 export * from '#x/BaseDataClass';
+export * from '#x/ByteCount';
 export * from '#x/Converter';
 export * from '#x/ConverterConfig';
 export * from '#x/Duration';